Update README.md
I hope this message finds you well. I recently submitted a pull request (PR) to your project as part of Hacktoberfest, and I would be grateful if you could take a moment to review it.

If there are any changes or improvements needed, please feel free to let me know, and I would be happy to address them.

Thank you for maintaining such a great project, and I look forward to your feedback.